requirements.txt
tornado==6.4.1
web3==6.20.1
eth-account==0.13.4
python-dotenv==1.0.1
